+++ .gitignore
@@ -285,3 +285,4 @@ $RECYCLE.BIN/
 *.msm
 *.msp
 /Input/
+/.idea/
